requirements.txt
apache-airflow-providers-amazon
sqlparse
sql-metadata
unittestreport
coverage
boto3